+++ src/services/oni-save/components/AbstractGeyserEditor/dispatch-props.ts
@@ -1,0 +1,34 @@
+import { Dispatch } from "redux";
+
+import {
+  modifyBehavior,
+  BehaviorDataTarget
+} from "../../actions/modify-behavior";
+import { changeGeyserType } from "../../actions/change-geyser-type";
+
+import { AbstractGeyserEditorProps } from "./props";
+import { GeyserBehavior } from "oni-save-parser";
+
+export default function mapDispatchToProps(
+  dispatch: Dispatch,
+  props: AbstractGeyserEditorProps
+) {
+  return {
+    onChangeEmitRate: (value: number) =>
+      dispatch(
+        modifyBehavior(
+          props.gameObjectId,
+          GeyserBehavior,
+          BehaviorDataTarget.Template,
+          {
+            configuration: {
+              rateRoll: value
+            }
+          }
+        )
+      ),
+    onChangeGeyserType: (type: string) =>
+      dispatch(changeGeyserType(props.gameObjectId, type))
+  };
+}
+export type DispatchProps = ReturnType<typeof mapDispatchToProps>;