+++ src/services/oni-save/components/AbstractGeyserEditor/index.ts
@@ -1,0 +1,2 @@
+import AbstractGeyserEditor from "./AbstractGeyserEditor";
+export default AbstractGeyserEditor;