+++ src/services/oni-save/components/AbstractGeyserEditor/props.ts
@@ -1,0 +1,10 @@
+export interface AbstractGeyserEditorProps {
+  gameObjectId: number;
+  children(props: AbstractGeyserEditorRenderProps): React.ReactChild;
+}
+export interface AbstractGeyserEditorRenderProps {
+  geyserType: string | null;
+  emitRate: number | null;
+  onChangeEmitRate(value: number): void;
+  onChangeGeyserType(type: string): void;
+}